requirements.txt
celery==5.2.3
coreapi==2.3.3
coreschema==0.0.4
django==4.0.1
django-celery-results==2.2.0
django-environ==0.8.1
django-filter==21.1
django-fsm==2.8.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==4.8.0
packaging==21.3
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
wrapt==1.13.3
uuid==1.30
drf-writable-nested==0.6.3
boto3==1.21.40
python-magic==0.4.25
fastapi==0.75.1
uvicorn==0.17.6
pydantic==1.9.0
pika==1.2.0
rabbitmq==0.2.0
sqlalchemy==1.4.43
django-cron==0.6.0